- created features package - created interfaces and viewmodels - renamed references
