+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.RandomUserApi"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".features.selecnumberofusers.selectnumberofusersactivity.SelectNumberOfUsersActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
