+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersactivity/SelectNumberOfUsersActivityInterface.kt
@@ -1,0 +1,4 @@
+package com.example.randomuserapi.features.selecnumberofusers.selectnumberofusersactivity
+
+interface SelectNumberOfUsersActivityInterface {
+}
