+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersviewmodel/SelectNumberOfUsersViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.randomuserapi.features.selecnumberofusers.selectnumberofusersviewmodel
+
+class SelectNumberOfUsersViewModel {
+}
