+++ app/src/main/java/com/example/randomuserapi/features/selecnumberofusers/selectnumberofusersviewmodel/SelectNumberOfUsersViewModelInterface.kt
@@ -1,0 +1,4 @@
+package com.example.randomuserapi.features.selecnumberofusers.selectnumberofusersviewmodel
+
+interface SelectNumberOfUsersViewModelInterface {
+}
