+++ app/src/main/res/layout/activity_select_number_of_users.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".features.selecnumberofusers.selectnumberofusersactivity.SelectNumberOfUsersActivity">
 
     <TextView
         android:layout_width="wrap_content"
